fix spurious errors while compiling to pdf
The makepdf script monitors latex by looking for lines beginning
by '! ' in the logs. Unfortunately, it uses a regexp while latex
seems to spit lines which are not perfect utf-8. The regexp may raise
an exception.

It is safer to just check the first two characters of each line.
This may fix github/progit#12