Make two timestamps in diff output more distinct.
It takes time to scan over two dates in the output to find out what is
the difference between them. And then a reader may wonder why the
new version is older then the previous one. So make two timestamps
more distinct to avoid unnecessary eyeballing of the example output.
This change improves readability a bit.