08-git-and-other-scms: Fix command for svn author names
The command for creating the svn user list is wrong.
It does not output the list in the format which is required
by git-svn clone.
This issue has already been fixed for some languages in
commit 758d464c3d1d9ce31021c70ed2fd0c08d40772b3.
This patch applies this fix for all languages.
Additionally the output is directly written into the file
users.txt where it is expected to be according to the
subsequent git-svn clone command.