fixed error in the imap section
git send-email does not put the patches into an IMAP folder, but the
command git imap-send.
I've changed the command, added the output that occurs when using the
command and wrote a short sentence to make use of the previous output
used in the document. I've also added an example for SMTP config.